+++ package.json
@@ -28,6 +28,7 @@
     "@hookform/resolvers": "^3.1.1",
     "ace-builds": "^1.36.5",
     "classnames": "^2.3.2",
+    "file-loader": "^6.2.0",
     "fuzzy": "^0.1.3",
     "react-ace": "^13.0.0",
     "react-hook-form": "^7.45.1",
