+++ packages/esm-billing-app/src/billable-services/billable-exemptions/schema-editor/schema-editor.scss
@@ -1,23 +1,24 @@
 @use '@carbon/colors';
 @use '@carbon/type';
+@use '@carbon/layout';
 
 .validationErrorsContainer {
   display: flex;
-  gap: 0.5rem;
+  gap: layout.$spacing-03;
 
   :global(.cds--actionable-notification) {
-    max-width: 38rem;
+    max-width: 100%;
   }
   :global(.cds--actionable-notification__details) {
-    margin: 0 0rem 0 0.8125rem !important;
+    margin: 0 0rem 0 layout.$spacing-04 !important;
   }
 }
 
 .pagination {
   display: flex;
   align-items: center;
-  gap: 0.5rem;
-  font-size: 0.875rem;
+  gap: layout.$spacing-03;
+  font-size: layout.$spacing-04;
 
   .paginationIcon {
     fill: colors.$blue-50;
@@ -28,4 +29,8 @@
     fill: colors.$gray-40;
     cursor: not-allowed;
   }
+}
+
+.hasErrors {
+  border: layout.$spacing-02 solid colors.$red-60;
 }
