+++ packages/esm-billing-app/src/hooks/useExemptionSchema.tsx
@@ -33,3 +33,17 @@ export async function saveExemptionSchema(schema: Schema, key: string): Promise<
     body,
   });
 }
+
+export function useStandardSchema(key: string) {
+  const { data, error, isLoading } = useSWRImmutable<{ data }, Error>(
+    `/ws/rest/v1/systemsetting?q=${key}&v=full`,
+    openmrsFetch,
+  );
+
+  return {
+    schema: data?.data,
+    schemaProperties: data?.data?.results?.find((resource) => resource.property === 'kenyaemr.billing.exemptions'),
+    error,
+    isLoading,
+  };
+}
